backend/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
google-genai
python-dotenv
firecrawl-py
requests
beautifulsoup4
click
groq
